requirements.txt
anthropic
google-genai
python-dotenv
fastapi
uvicorn
libsql-experimental
markdown
python-docx
openai
numpy
rich
tqdm

Pillow
